+++ Source/Noise.h
@@ -9,17 +9,17 @@
 */
 
 #pragma once
-
+#include <stdint.h>
 double uniform_closed(double a, double b);
 
 double deterministic_uniform_closed(double a, double b, double key);
 
-double coin_flip();
+int64_t coin_flip();
 
 double valueNoise(double x);
 
 double voronoiNoise(double x);
 
-double perlinNoise(double x, int octaveCount, double lacunarity, double roughness);
+double perlinNoise(double x, int64_t octaveCount, double lacunarity, double roughness);
 
-double ridgedMultiNoise(double x, int octaveCount, double lacunarity);
+double ridgedMultiNoise(double x, int64_t octaveCount, double lacunarity);
